Prints table of diversities

Print the frequency table of diversities from a hagis.diversities object The resulting object is a pander table (a text object for Markdown) for ease of use in reporting and viewing in the console.

diversities_table(x, ...)

Arguments

x
a hagis.diversities object generated by calculate_diversities()

...
other arguments passed to pander::panderOptions()

Value

A pander object of diversities

See also

Examples

# Using the built-in data set, P_sojae_survey
data(P_sojae_survey)

P_sojae_survey

#>      Isolate         Line         Rps Total HR (1) Lesion (2)
#>   1:       1     Williams susceptible    10      0          0
#>   2:       1       Harlon      Rps 1a    10      4          0
#>   3:       1 Harosoy 13xx      Rps 1b     8      0          0
#>   4:       1     L75-3735      Rps 1c    10     10          0
#>   5:       1    PI 103091      Rps 1d     9      2          0
#>  ---                                                         
#> 290:      21   PRX 145-48      Rps 3c     8      3          1
#> 291:      21     L85-2352       Rps 4    10      3          1
#> 292:      21     L85-3059       Rps 5    10      0          4
#> 293:      21 Harosoy 62XX       Rps 6     8      2          1
#> 294:      21      Harosoy       Rps 7    10      0          0
#>      Lesion to cotyledon (3) Dead (4) total.susc total.resis perc.susc
#>   1:                       0       10         10           0       100
#>   2:                       0        6          6           4        60
#>   3:                       0        8          8           0       100
#>   4:                       0        0          0          10         0
#>   5:                       1        6          7           2        78
#>  ---                                                                  
#> 290:                       0        4          5           3        63
#> 291:                       4        2          7           3        70
#> 292:                       0        6         10           0       100
#> 293:                       0        5          6           2        75
#> 294:                       0       10         10           0       100
#>      perc.resis
#>   1:          0
#>   2:         40
#>   3:          0
#>   4:        100
#>   5:         22
#>  ---           
#> 290:         38
#> 291:         30
#> 292:          0
#> 293:         25
#> 294:          0

# calculate susceptibilities with a 60 % cutoff value
diversities <- calculate_diversities(x = P_sojae_survey,
                                     cutoff = 60,
                                     control = "susceptible",
                                     sample = "Isolate",
                                     gene = "Rps",
                                     perc_susc = "perc.susc")

# print the diversities table
diversities_table(diversities)

#> 
#> --------------------------------------------
#>  Frequency             Pathotype            
#> ----------- --------------------------------
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3a, Rps 
#>              3b, Rps 3c, Rps 4, Rps 5, Rps  
#>                         6, Rps 7            
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3a, Rps 
#>                 3b, Rps 3c, Rps 5, Rps 7    
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3a, Rps 
#>                 3b, Rps 3c, Rps 6, Rps 7    
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3b, Rps 
#>                  4, Rps 5, Rps 6, Rps 7     
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3b, Rps 
#>                     4, Rps 5, Rps 7         
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3b, Rps 
#>                     4, Rps 6, Rps 7         
#> 
#>      2        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 3b, Rps 
#>                            7                
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1d, Rps 1k, Rps 2, Rps 6, Rps  
#>                            7                
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>               1d, Rps 1k, Rps 3a, Rps 3b,   
#>                   Rps 5, Rps 6, Rps 7       
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>                1d, Rps 1k, Rps 3b, Rps 7    
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1k, Rps 2, Rps 3b, Rps 3c, Rps 
#>                     4, Rps 6, Rps 7         
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>              1k, Rps 3b, Rps 5, Rps 6, Rps  
#>                            7                
#> 
#>      1        Rps 1a, Rps 1b, Rps 1c, Rps   
#>                 1k, Rps 3b, Rps 5, Rps 7    
#> 
#>      1        Rps 1a, Rps 1b, Rps 1d, Rps   
#>              1k, Rps 2, Rps 3a, Rps 3b, Rps 
#>                     5, Rps 6, Rps 7         
#> 
#>      2        Rps 1a, Rps 1c, Rps 1d, Rps   
#>              1k, Rps 2, Rps 3b, Rps 5, Rps  
#>                            7                
#> 
#>      1        Rps 1a, Rps 1c, Rps 1d, Rps   
#>              1k, Rps 2, Rps 3b, Rps 6, Rps  
#>                            7                
#> 
#>      1        Rps 1a, Rps 1c, Rps 1d, Rps   
#>                     3b, Rps 5, Rps 7        
#> 
#>      1       Rps 1a, Rps 1c, Rps 3b, Rps 5, 
#>                       Rps 6, Rps 7          
#> 
#>      1       Rps 1a, Rps 1c, Rps 3b, Rps 5, 
#>                          Rps 7              
#> --------------------------------------------
#>