hagis: Analysis of Plant Pathogen Pathotype Complexities, Distributions and Diversity

Analysis of plant pathogen pathotype survey data. Functions provided calculate distribution of susceptibilities, distribution of complexities with statistics, pathotype frequency distribution, as well as diversity indices for pathotypes. This package is meant to be a direct replacement for Herrmann, Löwer and Schachtel's (1999) doi:10.1046/j.1365-3059.1999.00325.x Habgood-Gilmour Spreadsheet, 'HaGiS', previously used for pathotype analysis.

See also

Author

Maintainer: Adam H. Sparks adamhsparks@gmail.com (ORCID)

Authors:

Austin G. McCoy mccoyaus@msu.edu (ORCID) [conceptor]

Zachary Noel noelzach@msu.edu (ORCID) [conceptor]

Martin Chilvers chilvers@msu.edu (ORCID) [thesis advisor]

Jari Oksanen jhoksane@gmail.com (Contributed fix for the use of 'vegan' in betadiversity vignette.)

Other contributors:

Zhian N. Kamvar (ORCID) [contributor, reviewer]

Michigan Soybean Promotion Committee [funder]

Project GREEEN [funder]

North Central Soybean Research Program [funder]

GRDC Project DAQ00186 [funder]