Phytophthora sojae survey example data

Data from a Phytophthora sojae survey

data(P_sojae_survey)

Format

An object of class data.table with 294 observations of 12 variables

Isolate

P. sojae isolate identifier

Line

Soybean cultivar

Rps

Rps gene identifier

Total

Total number of plants inoculated

HR (1)

Number of plants that exhibit a hypersensitive response after inoculation

Lesion (2)

Number of plants that develop a lesion at inoculation site

Lesion to cotyledon (3)

Number of plants that develop a lesion, which advances to the hypocotyl of the seedling after infection

Dead (4)

Number of dead plants that are observed after inoculation

total.susc

The total number of susceptible plants (Lesion+Lesion to cotyledon+Dead)

total.resis

The total number of resistant plants (equal to HR value)

Source

Data from an ongoing 2017 Phytophthora sojae survey in Michigan, conducted by A. G. McCoy et al..

Examples

data(P_sojae_survey)
P_sojae_survey

#>      Isolate         Line         Rps Total HR (1) Lesion (2)
#>   1:       1     Williams susceptible    10      0          0
#>   2:       1       Harlon      Rps 1a    10      4          0
#>   3:       1 Harosoy 13xx      Rps 1b     8      0          0
#>   4:       1     L75-3735      Rps 1c    10     10          0
#>   5:       1    PI 103091      Rps 1d     9      2          0
#>  ---                                                         
#> 290:      21   PRX 145-48      Rps 3c     8      3          1
#> 291:      21     L85-2352       Rps 4    10      3          1
#> 292:      21     L85-3059       Rps 5    10      0          4
#> 293:      21 Harosoy 62XX       Rps 6     8      2          1
#> 294:      21      Harosoy       Rps 7    10      0          0
#>      Lesion to cotyledon (3) Dead (4) total.susc total.resis perc.susc
#>   1:                       0       10         10           0       100
#>   2:                       0        6          6           4        60
#>   3:                       0        8          8           0       100
#>   4:                       0        0          0          10         0
#>   5:                       1        6          7           2        78
#>  ---                                                                  
#> 290:                       0        4          5           3        63
#> 291:                       4        2          7           3        70
#> 292:                       0        6         10           0       100
#> 293:                       0        5          6           2        75
#> 294:                       0       10         10           0       100
#>      perc.resis
#>   1:          0
#>   2:         40
#>   3:          0
#>   4:        100
#>   5:         22
#>  ---           
#> 290:         38
#> 291:         30
#> 292:          0
#> 293:         25
#> 294:          0